Move epsilon removal algorithm into own module
